Add "archive_folders" to ImageBrowser

+++ modules/imagebrowser.py
@@ -6,9 +6,11 @@ import json
 from typing import Dict, List, Tuple, Optional
 from pathlib import Path
 import sqlite3
+import time
 from modules.path import PathManager
 from modules.util import TimeIt
 from modules.settings import default_settings
+import version
 
 
 def format_metadata(metadata: Dict) -> Dict:
@@ -138,9 +140,31 @@ def connect_database(path="cache/images.db"):
     # Connect to an SQLite database (or create it if it doesn't exist)
     conn = sqlite3.connect(path, check_same_thread=False)
 
-    cursor = conn.cursor()
-    # FIXME!!! need more columns for prompt and size etc
-    cursor.execute('''CREATE TABLE IF NOT EXISTS images (path text, json text)''')
+    conn.cursor()
+    conn.execute('''CREATE TABLE IF NOT EXISTS status (version text, date text)''')
+    res = conn.execute("SELECT count(*) FROM status")
+    cnt = res.fetchone()[0]
+    newdb = False
+    if cnt == 0:
+        conn.execute(
+            "INSERT INTO status(version, date) VALUES (?,?)",
+            (str(version.version), str(time.time()))
+        )
+        newdb = True
+    else:
+        res = conn.execute("SELECT version FROM status")
+        dbver = res.fetchone()[0]
+        if dbver != version.version:
+            newdb = True
+
+    if newdb:
+        try:
+            conn.execute("DROP TABLE images")
+        except:
+            pass
+    conn.commit()
+    conn.cursor()
+    conn.execute('''CREATE TABLE IF NOT EXISTS images (fullpath text, path text, json text)''')
     conn.commit()
 
     return conn
@@ -166,7 +190,7 @@ class ImageBrowser:
         if page == None:
             page = 1
         result = self.sql_conn.execute(
-            f"SELECT path FROM images WHERE json LIKE '%{self.filter}%' ORDER BY path DESC LIMIT ? OFFSET ?",
+            f"SELECT fullpath, path FROM images WHERE json LIKE '%{self.filter}%' ORDER BY path DESC LIMIT ? OFFSET ?",
             (
                 str(self.images_per_page),
                 str((page-1)*self.images_per_page),
@@ -175,47 +199,53 @@ class ImageBrowser:
         image_paths = result.fetchall()
         self.current_display_paths = image_paths  # Store current display order
         if image_paths:
-            path1 = str(Path(image_paths[0][0]).relative_to(self.base_path))
-            path2 = str(Path(image_paths[-1][0]).relative_to(self.base_path))
+            path1 = str(Path(image_paths[0][1]))
+            path2 = str(Path(image_paths[-1][1]))
         else:
             path1 = "None"
             path2 = "None"
         text = f"{path1} ... {path2}"
 
         if image_paths:
-            return list(zip(*image_paths))[0], text
+            return list(Path(x[0]) for x in image_paths), text
         return [], text
 
     def update_images(self) -> Tuple[List[str], str]:
         """Check all images and update database"""
-        try:
+        #try:
+        if True:
             if not self.base_path.exists():
                 return [], f"Folder not found: {self.base_path}"
 
             image_cnt = 0
+            self.sql_conn.cursor()
             self.sql_conn.execute("DROP TABLE images")
-            self.sql_conn = connect_database() # Re-connect and re-create database
+            self.sql_conn.commit()
+            self.sql_conn = connect_database()
             self.sql_conn.cursor()
 
             # Walk through directory and all subdirectories
             with TimeIt("Update DB"):
-                for root, _, files in os.walk(self.base_path):
-                    for filename in files:
-                        #if filename.lower().endswith((".png", ".gif")):
-                        if filename.lower().endswith(".png"):
-                            full_path = Path(root) / filename
-                            rel_path = str(full_path.relative_to(self.base_path))
+                for folder in [self.base_path] + default_settings.get("archive_folders", []):
+                    print(f"DEBUG: {folder}")
+                    for root, _, files in os.walk(folder):
+                        print(f"DEBUG: {files}")
+                        for filename in files:
+                            #if filename.lower().endswith((".png", ".gif")):
                             if filename.lower().endswith(".png"):
-                                metadata = get_png_metadata(str(full_path))
-                            else:
-                                metadata = {} # FIXME fake data for non-png images
-                            metadata["file_path"] = rel_path
+                                full_path = Path(root) / filename
+                                rel_path = str(full_path.relative_to(folder))
+                                if filename.lower().endswith(".png"):
+                                    metadata = get_png_metadata(str(full_path))
+                                else:
+                                    metadata = {} # FIXME fake data for non-png images
+                                metadata["file_path"] = rel_path
 
-                            self.sql_conn.execute(
-                                "INSERT INTO images(path, json) VALUES (?,?)",
-                                (str(full_path), json.dumps(metadata))
-                            )
-                            image_cnt += 1
+                                self.sql_conn.execute(
+                                    "INSERT INTO images(fullpath, path, json) VALUES (?, ?,?)",
+                                    (str(full_path), str(rel_path), json.dumps(metadata))
+                                )
+                                image_cnt += 1
 
             self.sql_conn.commit()
 
@@ -236,18 +266,18 @@ class ImageBrowser:
                 gr.update(value=f"No images found in {self.base_path} or its subdirectories")
             )
 
-        except Exception as e:
-            return (
-                gr.update(value=["html/error.png"]),
-                gr.update(value=1, maximum=1),
-                gr.update(value=f"Error updating folder: {e}")
-            )
+        #except Exception as e:
+        #    return (
+        #        gr.update(value=["html/error.png"]),
+        #        gr.update(value=1, maximum=1),
+        #        gr.update(value=f"Error updating folder: {e}")
+        #    )
 
     def get_image_metadata(self, evt: gr.SelectData) -> str:
         """Get metadata for selected image."""
         try:
-            selected_path = self.current_display_paths[evt.index]
-            result = self.sql_conn.execute("SELECT json FROM images WHERE path = ?", selected_path)
+            selected_path = self.current_display_paths[evt.index][0]
+            result = self.sql_conn.execute("SELECT json FROM images WHERE fullpath = ?", (str(selected_path),))
             data = json.loads(result.fetchone()[0])
             return format_metadata_string(data)
         except Exception as e:
